fix(web): Polish ApplicationReadyGuard
