+++ apps/web/src/ApplicationReadyGuard.tsx
@@ -1,0 +1,19 @@
+import { type PropsWithChildren, useLayoutEffect } from 'react';
+import { useAuth } from '@novu/shared-web';
+
+export function ApplicationReadyGuard({ children }: PropsWithChildren<{}>) {
+  const { isLoading, inPublicRoute } = useAuth();
+
+  // Clean up the skeleton loader when the app is ready
+  useLayoutEffect(() => {
+    if (inPublicRoute || !isLoading) {
+      document.getElementById('loader')?.remove();
+    }
+  }, [inPublicRoute, isLoading]);
+
+  if (inPublicRoute || !isLoading) {
+    return <>{children}</>;
+  }
+
+  return null;
+}
